Change default font in Plan 9
"lucidasans" Plan 9 font is called "lucsans" in plan9port
as part of an agreement with Bigelow & Holmes.
See https://marc.info/?l=9fans&m=114412454010468&w=2

Helps #115
